requirements.txt
# Recipe-to-Cart dependencies
# Python 3.10+

# Data and numerics
pandas>=2.0.0
numpy>=1.24.0

# Exploration and evaluation
matplotlib>=3.7.0
seaborn>=0.12.0

# NLP / retrieval
scikit-learn>=1.3.0
rank-bm25>=0.2.2

# Ollama HTTP client
requests>=2.28.0
# Optional: spaCy for fallback parsing
# spacy>=3.7.0
# nltk>=3.8.0

# Config and paths
pyyaml>=6.0.0

# Testing
pytest>=7.0.0
